% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TwoDimensional.R
\name{Irregular2dWavelet}
\alias{Irregular2dWavelet}
\title{Setting wavelet coefficients to zero that correspond to certain levels}
\usage{
Irregular2dWavelet(x, y, f, numLevels)
}
\arguments{
\item{x}{Vector of locations of the data in the first direction}

\item{y}{Vector of locations of the data in the second direction}

\item{f}{Vector of data values observed at the given locations}

\item{numLevels}{Number of wavelet levels. Should be an integer between 2 and 10}
}
\value{
A matrix of wavelet basis functions evaluated at the data as well as the estimated wavelet coefficients
}
\description{
Takes in a vector of coefficients and returns the same vector, but with entries forced to zero that correspond
to the desired wavelet levels
}

