% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TwoDimensional.R
\name{threshold2d}
\alias{threshold2d}
\title{Setting wavelet coefficients to zero that correspond to certain levels}
\usage{
threshold2d(beta, numLevels, remove.x, remove.y)
}
\arguments{
\item{beta}{Vector of wavelet coefficients}

\item{numLevels}{Number of wavelet levels. Should be an integer between 2 and 10}

\item{remove.x}{Vector of wavelet levels in the X direction that you want removed. Should contain
numbers between 1 and numLevels}

\item{remove.y}{Vector of wavelet levels in the Y direction that you want removed. Should contain
                      numbers between 1 and numLevels}
}
\value{
A vector of wavelet coefficients with those corresponding to basis functions involving levels in
        remove.x or remove.y set to zero
}
\description{
Takes in a vector of coefficients and returns the same vector, but with entries forced to zero that correspond
to the desired wavelet levels
}
\examples{
n <- 1000
x <- runif(n)
y <- runif(n)

numLevels <- 3
k = 2^numLevels - 1
Zx <- cbind(rep(1, n), ZDaub(x, numLevels=numLevels))
Zy <- cbind(rep(1, n), ZDaub(y, numLevels=numLevels))
Zxy <- Z2d(Zx,Zy)
}

