% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoDimensional.R
\name{Z2d}
\alias{Z2d}
\title{Creating design matrix with 2d wavelet basis functions as predictor}
\usage{
Z2d(Zx, Zy)
}
\arguments{
\item{Zx}{An N x K matrix where the columns are the 1d wavelet basis functions in the X direction
evaluated at the N locations of the data}

\item{Zy}{An N x K matrix where the columns are the 1d wavelet basis functions in the Y direction
evaluated at the N locations of the data}
}
\value{
An N x K^2 matrix containing products of wavelet basis functions
}
\description{
Takes in two matrices: The N x K matrices generated from 1d wavelet basis functions in each direction
}
\examples{

n <- 1000
x <- runif(n)
y <- runif(n)

numLevels <- 3
k = 2^numLevels - 1
Zx <- cbind(rep(1, n), ZDaub(x, numLevels=numLevels))
Zy <- cbind(rep(1, n), ZDaub(y, numLevels=numLevels))
Zxy <- Z2d(Zx,Zy)
}

