% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoDimensional.R
\name{Irregular2dWavelet}
\alias{Irregular2dWavelet}
\title{Performing 2d wavelet decomposition}
\usage{
Irregular2dWavelet(x, y, f, numLevels)
}
\arguments{
\item{x}{Vector of locations of the data in the first direction}

\item{y}{Vector of locations of the data in the second direction}

\item{f}{Vector of data values observed at the given locations}

\item{numLevels}{Number of wavelet levels. Should be an integer between 2 and 10}
}
\value{
A matrix of wavelet basis functions evaluated at the data as well as the estimated wavelet coefficients
}
\description{
Takes in the location of the data as well as the data values
}
\examples{

data(BostonPMsim)
numLevels <- 3
Decomp <- Irregular2dWavelet(x=BostonPMsim$long, 
                             y=BostonPMsim$lat,
                             f=BostonPMsim$pm,
                             numLevels=numLevels)

threshold.betalow4 <- threshold2d(Decomp$beta, numLevels, remove.x=4:numLevels, remove.y=4:numLevels)
threshold.fHatlow4 <- Decomp$Zxy\%*\%threshold.betalow4

threshold.col = colors(100)[as.numeric(cut(threshold.fHatlow4,breaks = 100))]
map("state","massachusetts", xlim=range(BostonPMsim$long), ylim=range(BostonPMsim$lat))
points(BostonPMsim$long, BostonPMsim$lat, col=threshold.col, cex=.2)
}

