% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TwoDimensional.R
\name{Irregular2dWavelet}
\alias{Irregular2dWavelet}
\title{Performing 2d wavelet decomposition}
\usage{
Irregular2dWavelet(x, y, f, numLevels)
}
\arguments{
\item{x}{Vector of locations of the data in the first direction}

\item{y}{Vector of locations of the data in the second direction}

\item{f}{Vector of data values observed at the given locations}

\item{numLevels}{Number of wavelet levels. Should be an integer between 2 and 10}
}
\value{
A matrix of wavelet basis functions evaluated at the data as well as the estimated wavelet coefficients
}
\description{
Takes in the location of the data as well as the data values
}
\examples{
n <- 1000
x <- runif(n)
y <- runif(n)

numLevels <- 3
k = 2^numLevels - 1
Zx <- cbind(rep(1, n), ZDaub(x, numLevels=numLevels))
Zy <- cbind(rep(1, n), ZDaub(y, numLevels=numLevels))
Zxy <- Z2d(Zx,Zy)
}

