% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoDimensional.R
\name{threshold2d}
\alias{threshold2d}
\title{Setting wavelet coefficients to zero that correspond to certain levels}
\usage{
threshold2d(beta, numLevels, remove.x, remove.y)
}
\arguments{
\item{beta}{Vector of wavelet coefficients}

\item{numLevels}{Number of wavelet levels. Should be an integer between 2 and 10}

\item{remove.x}{Vector of wavelet levels in the X direction that you want removed. Should contain
numbers between 1 and numLevels}

\item{remove.y}{Vector of wavelet levels in the Y direction that you want removed. Should contain
numbers between 1 and numLevels}
}
\value{
A vector of wavelet coefficients with those corresponding to basis functions involving levels in
        remove.x or remove.y set to zero
}
\description{
Takes in a vector of coefficients and returns the same vector, but with entries forced to zero that correspond
to the desired wavelet levels
}
\examples{

data(BostonPMsim)
numLevels <- 3
Decomp <- Irregular2dWavelet(x=BostonPMsim$long, 
                             y=BostonPMsim$lat,
                             f=BostonPMsim$pm,
                             numLevels=numLevels)

threshold.betalow4 <- threshold2d(Decomp$beta, numLevels, remove.x=4:numLevels, remove.y=4:numLevels)
threshold.fHatlow4 <- Decomp$Zxy\%*\%threshold.betalow4

threshold.col = colors(100)[as.numeric(cut(threshold.fHatlow4,breaks = 100))]
map("state","massachusetts", xlim=range(BostonPMsim$long), ylim=range(BostonPMsim$lat))
points(BostonPMsim$long, BostonPMsim$lat, col=threshold.col, cex=.2)
}

